Manage the report building through a class object
The commits creates a report_class class and makes the old report variable
an object of this new class.

The new approach makes the code much easier to manage.
Main improvements:

- all inside_* variables are now stored in a unique dictionary in the
object, whose variables are automatically set or unset when tags are
opened or closed;

- the dictionary of the inside_variables is accessible through the new
is_inside(tag) member function;

- no more open_* and close_* functions: they have all been replaced by two
unique member functions: open(tag, category="", title="") and close(tag);

- thanks to the replacements of the inside_*, open_* and close_* variables
it is now possible to act on a list of tags through a for loop (and I
already used this possibility);

- functions that act directly on the report text are now clearly separated
from the other functions.